+++ src/lib.rs
@@ -24,6 +24,7 @@ mod point;
 mod rand;
 mod rect;
 mod tick;
+mod util;
 #[cfg(target_arch = "wasm32")]
 mod web;
 #[cfg(not(target_arch = "wasm32"))]